requirements.txt
langchain
langchain-openai
langchain-mongodb
langchain-text-splitters
pymongo
python-dotenv
certifi